Client

Description

Core module that handles all major game processes.

Game launch related functions

launchGame :: IO ()

Launch game screen.

parsePlaystyle :: IO [String] -> IO PlayStyleType

Parse command line arguments.

initTableScreen :: StdGen -> PlayStyleType -> IO TableScreen

Initialization of table screen. All images are loaded and all player data is set.

createTableScreenWith :: StdGen -> PlayStyleType -> TableImages -> TableScreen

Create new table screen made of images and set default parameters.

Game maintance related functions

updateGame :: Float -> TableScreen -> TableScreen

Update game parameters depending on game state.