Poker.AI.Engine

Description

Here is located all functions related to AI computations.

Core functions for simulating AI thinking

calculateAIMove :: Player -> Int -> Int -> Int -> Street -> (Move, AIPlayer)

Main function that return AI move and updated AI data depending on it's play style, incoming bet, blind size, pot and street.

calculatePreFlopMove :: [Card] -> PlayStyle -> Int -> Int -> Int -> Int -> Move

Calculate move on preflop depending on hand, play style, made bet, incoming bet, max bet and blind size.

calculatePostFlopMove :: Combination -> PlayStyle -> Int -> Int -> Int -> Int -> Street -> Bool -> StdGen -> (Move, StdGen)

Calculate move on postflop depending on combination, play style, made bet, incoming bet, max bet, pot size, street and last move.

Utility functions

updateAIData :: Player -> [Card] -> Player

Update AIPlayer data for player.

writeAIDataChange :: [Player] -> Position -> AIPlayer -> [Player]

Write changes to AI data to player on given position.

checkChance :: Int -> StdGen -> (Bool, StdGen)

Return if chance in range 0-100 succeed.

checkRange :: [Card] -> CardRange -> Bool

Return if hand is in range.

evalBet :: Int -> BetRange -> BetType

Return type of bet.

evalHand :: Combination -> CombinationRange -> HandPower

Return strength of combination.

suggestPFMove :: [Card] -> HandRangePF -> ActionType

Return suggested move for hand on preflop.