Poker.Interface.Handlers

Description

Module that operates with input events.

Handler functions

handleInput :: Event -> TableScreen -> TableScreen

Operate with user input.

handleClick :: (Float, Float) -> TableScreen -> TableScreen

Handle mouse click.

handleSmallButtonClick :: Int -> Int -> Slider -> Slider

Handle small button click depending on button number and pot size.

handleSliderClick :: Float -> Int -> Slider -> Slider

Handle slider click depending on click horizontal position.

moveSliderBall :: Float -> Int -> Slider -> Slider

Moves slider ball if possible.

Utility functions

updateSlideData :: Slider -> Player -> Int -> Int -> Slider

Updates slider min and max values depending on player and incoming bet.

checkButtonHit :: (Float, Float) -> ((Float, Float), (Float, Float)) -> Bool -> Int

Return clicked button number depending on first button bounding rectangle.

checkSliderHit :: (Float, Float) -> Bool

Return if clicked slider area.

Constants

buttonHitbox :: ((Float, Float), (Float, Float))

Bounding rectangle for first button.

smallButtonHeight :: Float

Height of small button.

smallButtonHitbox :: ((Float, Float), (Float, Float))

Bounding rectangle for first small button.

sliderPadding :: (Float, Float)

Slider inside padding to slider ball area.

sliderDimensions :: (Float, Float)

Slider dimensions.

sliderHitbox :: ((Float, Float), (Float, Float))

Slider horizontal bounds.

clickTime :: Float

Time to show click on button.