Poker.Logic.Calculations

Description

Module to deal with computation related to poker combinations.

Functions to operate with cards

checkStraight :: [Card] -> (Bool, CardRank)

Return if card array is a straight and it's rank.

checkFlush :: [Card] -> (Bool, [CardRank])

Return if card array is a flush and 5-7 cards ranks.

computeHandRank :: [Card] -> (HandRank, ([CardRank], [CardRank]))

Return hand rank and kicker cards depending on card array.

computeCombination :: Maybe (Card, Card) -> [Card] -> Combination

Compute combination from hand and board.

Utility functions

cardToStraightNumList :: [Card] -> [Int]

Convert card list to sort int list with removed duplicates, Ace card is counted twice for A2345 straights.

getSequenceLength :: [Int] -> Int

Get length of sequence that starts from head element of list.

getSequencesLength :: [Int] -> [Int]

Convert int list to list that contain max sequence starting from each element.

countSuits :: [Card] -> [Int]

Count amount of each suit in card list.

countRanks :: [Card] -> [Int]

Count amount of each rank in card list.

takeEqualBestN :: Int -> Int -> [Int] -> [Int]

Return indexes of n last elements that are equal to given.

markWinningCombinations :: [(Bool, Combination)] -> [Bool]

Convert combination list to bool list with marked top combinations. Receive combination list zipped with bool list that indicates if the combination require to participate in comparing.