Poker.Logic.Trading

Description

Contains stuff to process bet rounds.

Operations with positions

getFirstPosition :: Int -> Street -> Position

Return first position depending on amount of players and street.

getNextPositon :: Int -> Street -> Position -> Position

Return next position depending on amount of players and street.

getNextPositionClockwise :: Position -> Int -> Position

Return next position by clock arrow depending on player amount.

getLastPosition :: Int -> Street -> Position

Return last position depending on amount of players and street.

getSeatOfPosition :: Position -> [Player] -> Seat

Return seat of player on given position. Unsafe function for [] and lists without player with given position.

Computations with player(-s)

checkSkipForActivePlayer :: Player -> Int -> Int -> Bool

Check if active player is skippable depending on max bet and amount of players in hand that aren't all-in.

countInHandPlayers :: [Player] -> Int

Return amount of players left in hand.

countCanMovePlayers :: [Player] -> Int

Return amount of players in hand that can require move.

countMaxBet :: [Player] -> Int

Return maximal bet that occured.

checkReTrade :: [Player] -> Int -> Bool

Return if repeating of trade is needed.

checkGameEnd :: [Player] -> Bool

Return if game ended.

checkWin :: [Player] -> String -> Bool

Return if player with given name won the game.

autoHumanMove :: Player -> Int -> Move

Return default move to proposed bet size when human didn't made any input.

calculatePot :: [Player] -> Int

Calculate pot.

getMoveFromButtonPressed :: Int -> (ActionType, ActionType) -> Int -> Int -> Player -> Move

Return move depending on pressed button number, possible actions, incoming bet, selected bet, made bet and player balance.

Operations with player(-s)

getActivePlayer :: [Player] -> Player

Return active player. Isn't safe for [] case.

toggleNewActivePlayer :: [Player] -> Position -> [Player]

Set active player depending to given position and unsets previous active player.

writeMove :: [Player] -> Position -> Move -> [Player]

Write move to player on given position.

applyMoveResults :: [Player] -> [Player]

Decrease balance by bet, increase invested by bet, vanish move if can.

computeHandResults :: [Player] -> [Card] -> [Player]

Find out hand results and reward winner(-s) or split pot if draw. Also open all cards that should be shown at showdown. Vanish AI data relative to hand.

giveLeftPart :: Int -> [Player] -> [Player]

Give award to first active player.

markWinnersActive :: [Player] -> [Card] -> (Int, [Player])

Return amount of winners among player that invested something and mark them as active.

takePotFromPlayer :: Player -> Int -> (Player, Int)

Take from player part of invested sized in pot.

takePotFromPlayers :: [Player] -> Int -> ([Player], Int)

Take from each player part of invested sized in pot.

changePlayerPositions :: [Player] -> [Player]

Change player positions.

getPossibleActions :: Player -> Int -> (ActionType, ActionType)

Get possible actions for player depending on incoming bet. Return only second two actions. Fold action is calculated independently.

writeButtonClick :: Int -> [Player] -> [Player]

Write number of pressed button to active player.

Constants

aiThinkTime :: Float

Time to get response from AI player.

humanThinkTime :: Float

Time to get response from human player.

showdownTime :: Float

Time to display showdown results.